Droppable: Move droppable into widgets folder
